gh-76: De-automate extraction of the untokenized tail (#85)
*This PR effectively rolls back gh-80.*

A symbol alternative (like A | B | C | ...) in a root symbol of the
tokenizer grammar would naturally map to `enum` by pest-ast if not the
unparsed tail making the rule into (A | B | ..., tail).

Thus, to get our `enum` we need to either wrap the alternative-only root
into another, real root with the tail, or return the out-of-grammar
manual tail extraction. The former way does not work with future
multi-goal-symbol approach of ECMA-262 so we roll the change back
instead.