gh-9: Fix tail extraction attempt on parse error in get_next_token (#86)
`get_next_token` tries to get the unparsed tail by chopping of a
recognized token no matter whether we got the token or a parse error. As
a result, the error yields to a panic.

This PR moves the tail extraction into processing of parse success.
Also, it adds a regression test that uses `claims` crate.