gh-131: Fix grammar parameters in square brackets
"~" in the right side non-terminal parameter list means "this parameter
exists so we mention it for a possible linter, but do not use here".

In the fixed case of `ScriptBody :
StatementList[~Yield, ~Await, ~Return]` the parameter list means
"neither `yield`, nor `await`, nor `return` is allowed in the top-level
code, outside of functions".